arts/requirements.txt
Django==1.8
Pillow==2.8.1
PyMySQL==0.6.6
Unidecode==0.04.17
argparse==1.3.0
django-activelink==0.4
django-autocomplete-light==2.0.9
django-ckeditor==4.4.7
django-classy-tags==0.6.1
django-sekizai==0.8.1
django-simple-menu==1.0.9
django-tagging==0.3.4
django-uuslug==1.0.5
django-wysiwyg==0.7.0
pbr==0.10.8
python-slugify==1.0.2
six==1.9.0
sorl-thumbnail==12.2
stevedore==1.3.0
virtualenv==12.1.1
virtualenv-clone==0.2.5
virtualenvwrapper==4.5.0
